Don't send all non-printables.
e.g. shift on its own isn't very useful, and was being added to the text
in insert mode.

This shows a flaw in my testing technique: The browser event layer isn't
involved in the tests at all. Will need a headless browser at some point to
catch this kind of bug.