Remove hacky operation reset.
When replaying a macro, we don't want the macro's key presses to be
considered part of the "replay macro" operation. Initially I got around this
by manually starting a new operation, but this required knowledge of the
inner workings of `NormalMode`. This commit switches to using `setTimeout`
with a timeout of zero, which will replay the macro immediately after the
current event handler has finished.